glacial_lake_segmentation/requirements.txt
# ============================================================
# INSTALL ORDER (RTX 5 series GPU — CUDA 12.8):
# Step 1 — PyTorch nightly (MUST be installed first):
#   pip install --pre torch torchvision --index-url https://download.pytorch.org/whl/nightly/cu128
# Step 2 — Everything else:
#   pip install -r requirements.txt
# ============================================================

numpy
pillow
matplotlib
scikit-learn
tqdm
albumentations
opencv-python-headless
fastapi
uvicorn[standard]
python-multipart
aiofiles
